--- src/utilities/chunk.js
@@ -1,3 +1,0 @@
-export default function chunk(array, size) {
-    return array
-}
